+++ src/py/flwr/client/clientapp/clientappio_servicer_test.py
@@ -17,7 +17,7 @@
 import unittest
 from unittest.mock import Mock, patch
 
-from flwr.client.clientapp.process import pull_message, push_message
+from flwr.client.clientapp.app import pull_message, push_message
 from flwr.common import Context, Message, typing
 from flwr.common.serde import (
     clientappstatus_from_proto,
